(HP-1031) Extracted useApiAccessTokens hook to own file
Profile.ts should have only profile related stuff
